recipies/requirements.txt
django
djangorestframework
drf-spectacular
djangorestframework-simplejwt
celery
redis
Django==5.0
celery
redis
djangorestframework
Pillow
django-celery-beat
boto3